+++ pkg/vm_service/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 15.2.0-wip
+- Update to version `4.21` of the spec.
+- Add `breakpointsUpdates` parameter to `reloadSources`.
+- Add `breakpointsUpdateReports` property to `ReloadReport`.
+- Add `BreakpointsUpdate`, `BreakpointLocation`, `BreakpointsUpdateReport` and
+  `BreakpointsUpdateFailure` types.
+
 ## 15.1.0
 - Update to version `4.20` of the spec.
 - Deprecate `streamCpuSamplesWithUserTag` RPC.